Create cond_var-synchronization.cpp
std::condition_variable is used to synchronize the threads.
The print_numbers thread waits for the ready flag to be true. It uses cv.wait(lock, condition) to block until the condition is met.
The set_ready thread simulates some work (delayed by 2 seconds) and then sets the ready flag to true and notifies the waiting thread with cv.notify_one().