Create thread_pool_queue.cpp
A simple implementation of a thread pool using std::thread and std::queue. Thread pools are useful when you want to manage a fixed number of threads that can process tasks concurrently.

This is a simple thread pool implementation.
The ThreadPool class manages a pool of worker threads that can process tasks in a queue.
Tasks (functions) are added to the queue via the enqueue method.
The worker threads continuously pull tasks from the queue and execute them. The pool shuts down when the ThreadPool object goes out of scope.